refactor: `SpaceBlobAddClient` is a `testutil`
